check if log file exists before rm

+++ shared.py
@@ -82,7 +82,7 @@ def get_res_bin_dir(ssa_algo: str, bin_name: str) -> Path:
     return get_res_dir(ssa_algo) / bin_name
 
 def get_decomp_file(ssa_algo: str, bin_name: str, func_name: str) -> Path:
-    return get_decomp_bin_dir(ssa_algo, bin_name) / func_name
+    return get_decomp_bin_dir(ssa_algo, bin_name) / func_namefi
 
 def get_src_func_file(ssa_algo: str, bin_name: str, func_name: str) -> Path:
     return get_src_func_bin_dir(ssa_algo, bin_name) / func_name
@@ -110,7 +110,8 @@ def clear_and_create_dir(path: Path, log_file: Path = LOG_FILE):
 
 def clear_log():
     try:
-        os.remove(LOG_FILE)
+        if os.path.exists(LOG_FILE):
+            os.remove(LOG_FILE)
     except Exception as e:
         print(f"Unable to remove log file {LOG_FILE}: {e}", file=sys.stderr)
